fifo: Add immutable group
ImmutableGroup is a martian.RequestResponseModifier that maintains lists of
request and response modifiers executed on a first-in, first-out basis.
ImmutableGroup can be constructed only from a fifo.Group.
Since it cannot be modified, it has no locks, and this gives a performance boost.

Group is now capable of transforming itself into ImmutableGroup.