Refactor tunneling code to tunnel functions
The purpose of this change is removing duplicated code and improving readability.

Both Proxy and proxyHandler add a tunnel function that is responsible for orchestrating copySync calls.
This allows to highlight both similarities between the two implementations and differences between handling CONNECT requests and Switching Protocols response.

In addition to that in proxyHandler the error propagation is fixed.
We only returned nil error.
The error handling in ServeHTTP has very limited use because it needs to be done prior to writing anything to the connection and hijacking.